+++ src/iters/iter.rs
@@ -59,40 +59,3 @@ impl<'a, K, V, const CAP: usize, I: PrimInt + Unsigned> DoubleEndedIterator
         Some(self.get_entry(i))
     }
 }
-
-/// Iterator that also returns the index of the current element
-///
-/// Used for internal implementation
-pub struct IterIndexed<'a, K, V, const CAP: usize, I: PrimInt + Unsigned>(Iter<'a, K, V, CAP, I>);
-
-impl<'a, K, V, const CAP: usize, I: PrimInt + Unsigned> IterIndexed<'a, K, V, CAP, I> {
-    pub fn new(const_lru: &'a ConstLru<K, V, CAP, I>) -> Self {
-        Self(Iter::new(const_lru))
-    }
-}
-
-impl<'a, K, V, const CAP: usize, I: PrimInt + Unsigned> Iterator for IterIndexed<'a, K, V, CAP, I> {
-    type Item = (I, &'a K, &'a V);
-
-    fn next(&mut self) -> Option<Self::Item> {
-        // consume then increment
-        let i = self.0.cursors.get_from_head();
-        self.0.next().map(|(k, v)| (i, k, v))
-    }
-
-    // TODO: look into https://doc.rust-lang.org/std/iter/trait.TrustedLen.html when it lands in stable
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        (0, Some(CAP))
-    }
-}
-
-impl<'a, K, V, const CAP: usize, I: PrimInt + Unsigned> DoubleEndedIterator
-    for IterIndexed<'a, K, V, CAP, I>
-{
-    fn next_back(&mut self) -> Option<Self::Item> {
-        // decrement then consume
-        self.0
-            .next_back()
-            .map(|(k, v)| (self.0.cursors.get_from_tail(), k, v))
-    }
-}
